# server/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_xinbreak_movie-web-app_internal_models.Video:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      views_count:
        type: integer
    type: object
...
  github_com_xinbreak_movie-web-app_internal_models_dtos.CreateVideoRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      duration:
        minimum: 0
        type: integer
      file_path:
        type: string
      thumbnail:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - file_path
    - title
    type: object
...
  github_com_xinbreak_movie-web-app_internal_models_dtos.UpdateVideoRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
  github_com_xinbreak_movie-web-app_internal_models_dtos.UserLoginDTO:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  github_com_xinbreak_movie-web-app_internal_models_dtos.UserRegisterDTO:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        minLength: 8
        type: string
      username:
        example: johndoe
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_xinbreak_movie-web-app_internal_models_dtos.UserResponseDTO:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  github_com_xinbreak_movie-web-app_internal_models_dtos.UserUpdateDTO:
    properties:
      avatar_url:
        type: string
      password:
        example: secret123
        type: string
      username:
        minLength: 3
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Movie Application
  title: movie-web-app API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserResponseDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserRegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /users/{id}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UserUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user profile
      tags:
      - users
  /users/{user_id}/videos:
    get:
      description: Get paginated list of videos for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get videos by user
      tags:
      - videos
  /videos:
    get:
      description: Get paginated list of all videos
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all videos
      tags:
      - videos
    post:
      consumes:
      - application/json
      description: Upload a new video with title and description
      parameters:
      - description: Video creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.CreateVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Video'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new video
      tags:
      - videos
  /videos/{id}:
    delete:
      description: Delete a video by ID (only owner can delete)
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a video
      tags:
      - videos
    get:
      description: Get a single video by its ID
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Video'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get video by ID
      tags:
      - videos
    put:
      consumes:
      - application/json
      description: Update an existing video (only owner can update)
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Video update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UpdateVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Video'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a video
      tags:
      - videos
...
swagger: "2.0"